requirements/requirements_local.txt
celery==4.1.0
django==1.11.9
requests>=2.18.4
simplejson>=3.13.0
djangorestframework=3.7.5
django-celery==3.2.2
redis>=2.10.0
django-debug-toolbar==1.9.1
django-cors==2.1.0